Data Managment Redone
NEEDS SANITY CHECK AND CLEAN UP
+ Remove all file-related things from Character
+ Redone JSONManager to use dictionary
+ Added Character deleting in BlazorUI

+++ Project20Core/JSONManager.cs
@@ -9,6 +9,30 @@ namespace Core
 {
     public class JSONManager
     {
+        private Dictionary<string, Character> _characters = new();
+
+        private const string EXTENSION = ".json";
+        private static string GetFilename(string id)
+        {
+            return $"{id}{EXTENSION}";
+        }
+
+        // TODO - Needs checking, also use it as minumum as possible
+        private bool TryGetCharacterID(Character character, out string? id)
+        {
+            if (_characters.ContainsValue(character))
+            {
+                id = _characters.FirstOrDefault(x => x.Value == character).Key;
+                return true;
+            }
+            else
+            {
+                id = null;
+                return false;
+            }
+        }
+
+        /*
         /// <summary>
         /// Deletes character's JSON.
         /// </summary>
@@ -20,10 +44,75 @@ namespace Core
             {
                 return;
             }
-            string filePath = Path.Combine(path, character.filename);
+            string filePath = Path.Combine(path, character.Filename);
+            File.Delete(filePath);
+        }
+        */
+
+        /// <summary>
+        /// Deletes character's JSON.
+        /// </summary>
+        /// <param name="character">Character that is to be removed.</param>
+        /// <param name="path">Path to the character folder.</param>
+        public void DeleteCharacter(Character character, string path)
+        {
+            if (!TryGetCharacterID(character, out string id)) return;
+
+            _characters.Remove(id);
+
+            if (!Directory.Exists(path))
+            {
+                return;
+            }
+
+            string filePath = Path.Combine(path, GetFilename(id));
             File.Delete(filePath);
         }
 
+        /// <summary>
+        /// Loads all character JSONs from given path.
+        /// <param name="path">Path to the character folder.</param>
+        /// </summary>
+        public Dictionary<string,Character> LoadCharacters(string path)
+        {
+            string[] filePaths;
+
+            // Checking if path exists, if not, creating one.
+            if (!Directory.Exists(path))
+            {
+                Directory.CreateDirectory(path);
+            }
+
+            filePaths = Directory.GetFiles(path);
+
+            foreach (var filePath in filePaths)
+            {
+                try
+                {
+                    using (StreamReader r = new StreamReader(filePath))
+                    {
+                        string json = r.ReadToEnd();
+
+                        Character? newCharacter = JsonSerializer.Deserialize<Character>(json);
+
+                        if (newCharacter == null) continue;
+
+                        string filename = Path.GetFileName(filePath);
+                        string id = filename.Remove(filename.Length - EXTENSION.Length); // Removing extension
+
+                        _characters[id] = (newCharacter);
+                    }
+                }
+                catch
+                {
+                    continue;
+                }
+            }
+
+            return _characters;
+        }
+
+        /*
         /// <summary>
         /// Loads all character JSONs from given path.
         /// <param name="path">Path to the character folder.</param>
@@ -52,7 +141,7 @@ namespace Core
                         if (newCharacter == null) continue;
 
                         characters.Add(newCharacter);
-                        newCharacter.filename = Path.GetFileName(filePath);
+                        newCharacter.Filename = Path.GetFileName(filePath);
                     }
                 }
                 catch
@@ -63,6 +152,7 @@ namespace Core
 
             return characters;
         }
+        */
 
         /// <summary>
         /// Loads all classes JSONs from given path.
@@ -142,6 +232,7 @@ namespace Core
             return races;
         }
 
+        /*
         /// <summary>
         /// Saves character to JSON to given path.
         /// </summary>
@@ -175,8 +266,48 @@ namespace Core
                 fileName = character.Name;
             }
 
-            character.filename = fileName + ".json";
-            string filePath = Path.Combine(path, character.filename);
+            character.Filename = fileName + ".json";
+            string filePath = Path.Combine(path, character.Filename);
+            File.WriteAllText(filePath, jsonString);
+        }
+        */
+
+        /// <summary>
+        /// Saves character to JSON to given path.
+        /// </summary>
+        /// <param name="character">Character that is to be saved.</param>
+        /// <param name="path">Path to the character folder.</param>>
+        /// <exception cref="ArgumentNullException"></exception>
+        public void SaveCharacter(Character character, string path)
+        {
+            string fileName;
+            string filePath;
+            string jsonString;
+
+            if (character == null)
+            {
+                throw new ArgumentNullException();
+            }
+
+            if (!Directory.Exists(path))
+            {
+                Directory.CreateDirectory(path);
+            }
+
+            jsonString = JsonSerializer.Serialize(character);
+
+            if (TryGetCharacterID(character, out string id))
+            {
+                fileName = id;
+            }
+            else
+            {
+                fileName = character.Name;
+                _characters[fileName] = character;
+            }
+
+            fileName = GetFilename(fileName);
+            filePath = Path.Combine(path, fileName);
             File.WriteAllText(filePath, jsonString);
         }
